+++ packages/gakumas-data/index.js
@@ -13,6 +13,12 @@ export {
   deserializePatchSequence,
 } from "./utils/effects";
 export { parseEffects, parsePatches } from "./utils/parser";
+export {
+  calculateMemoryStamina,
+  getAvailableTrueEndScenarios,
+  getMemoryStaminaBreakdown,
+  getSenseiStaminaBonus,
+} from "./utils/stamina";
 export { Tokenizer } from "./utils/parser/tokenizer";
 export { TokenType } from "./utils/parser/tokens";
 export { transformEffects, transformPatches } from "./utils/transformer";